msm: msm_bus: Fix the 64-bit division during interleaving
When interleaving flag is enabled, the bandwidth is divided
between ports. There was an error in the division as the
macro was not correctly updated with the right order of using
bandwidth and port number when 64-bit division function
was used by the bus driver.
Correcting this for fixing the token allocation issues
observed.

CRs-Fixed: 472791
Change-Id: Ie7c7f9557006cbb946eaa9eff4b50771cc60525e
Signed-off-by: Gagan Mac <gmac@codeaurora.org>
(cherry picked from commit 25a07480e9a25a75759d7d6db79b6f29350b8729)
Signed-off-by: caplio <github@snugplace.net>